+++ Packages/GameCanvas/Sample~/Acceleration/Assets/GameCanvas/AutoGeneratedScripts/GcFont.cs
@@ -1,0 +1,17 @@
+/*------------------------------------------------------------*/
+// <summary>GameCanvas for Unity</summary>
+// <author>Seibe TAKAHASHI</author>
+// <remarks>
+// (c) 2015-2020 Smart Device Programming.
+// This software is released under the MIT License.
+// http://opensource.org/licenses/mit-license.php
+// </remarks>
+/*------------------------------------------------------------*/
+namespace GameCanvas
+{
+    public readonly partial struct GcFont : System.IEquatable<GcFont>
+    {
+        internal const int __Length__ = 1;
+        public static readonly GcFont DefaultFont = new GcFont("GcFontDefaultFont");
+    }
+}